backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozendict==2.4.4
html5lib==1.1
idna==3.7
lxml==5.2.2
multitasking==0.0.11
mysqlclient==2.2.4
numpy==2.0.1
pandas==2.2.2
peewee==3.17.6
platformdirs==4.2.2
prometheus_client==0.20.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
yfinance==0.2.41
